feat: 프로젝트 페이지 모달 기능 구현
