+++ src/pages/Project.tsx
@@ -1,18 +1,32 @@
 import styled from '@emotion/styled'
-import React from 'react'
+import React, { useState } from 'react'
 import DefaultButton from '../components/DefaultButton'
 import ProjectList from '../components/ProjectList'
+import ProjectCreateModal from '../components/ProjectCreateModal'
 
 const Project = () => {
+
+  const [isModalOpen, setIsModalOpen] = useState(false)
+
+  const handleCreateProject = () => {
+    setIsModalOpen(true)
+    console.log('모달창 열림')
+  }
+
+  const handleModalClose = () => {
+    setIsModalOpen(false)
+  }
+
   return (
     <ProjectWrapper>
       <Title>내 프로젝트</Title>
       <Container>
         <BtnContainer>
-          <DefaultButton text='프로젝트 생성하기' width='9.2rem' height='2.5rem' fontSize='1rem' />
+          <DefaultButton text='프로젝트 생성하기' width='9.2rem' height='2.5rem' fontSize='1rem' onClick={handleCreateProject} />
         </BtnContainer>
         <ProjectList />
       </Container>
+      <ProjectCreateModal isOpen={isModalOpen} onClose={handleModalClose} />
     </ProjectWrapper>
   )
 }
